Allow decimal values for weight and waist measurements
Update schema definitions to use `real` for weight and waist columns in both `users` and `measurements` tables. Modify frontend logic in `profile-page.tsx` to use `parseFloat` when parsing weight values.

+++ client/src/pages/profile-page.tsx
@@ -386,7 +386,7 @@ export default function ProfilePage({ onClose }: ProfilePageProps) {
       const payload = {
         userId: user?.id,
         ...(data.weight &&
-          data.weight !== "" && { weight: parseInt(String(data.weight)) }),
+          data.weight !== "" && { weight: parseFloat(String(data.weight)) }),
         ...(data.waist &&
           data.waist !== "" && { waist: parseFloat(String(data.waist)) }),
       };
